// src/enums/unit-type.ts
export enum UsageUnitType {
    SIZE = 'SIZE',
    COUNT = 'COUNT',
}

export enum OverallUsageType {
    STORAGE = 'storage',
    CREDITS = 'credits',
    TOKENS = 'tokens',
}

export enum UnitType {
    BYTE = 'B',
    KILOBYTE = 'KB',
    MEGABYTE = 'MB',
    GIGABYTE = 'GB',
    TERABYTE = 'TB',
    DEFAULT = '',
}

export enum UnitPrefix {
    DEFAULT = 'DEFAULT',
    K = 'K',
    NONE = '',
    COUNT = 'COUNT',
    OUTPUT = 'OUTPUT',
}

export enum TimeUnitType {
    EMPTY = '',
    MINUTES = 'Minutes',
    HOURS = 'Hours',
    DAYS = 'Days',
}